Add humidity controller
This kicks in when we're about to shut off the fan to determine if it
would be beneficial to ingest outdoor air. If the indoor temperature
stays within range or the outdoor temperature is outside of range this
won't trigger.

This also doubles as a hack to reduce changing our on/off state which
can lead to desyncs.