Turn off drying if below cooling threshold
Right now we can replace "cool" with "dry" when we're both too hot & too
humid. We'll freely exchange between the two while sitting above the
cooling threshold. However, once we cross the cooling threshold we'll
stick to "dry" forever.

This adds a condition where if we're below the humidity midpoint (40%)
and drying we'll switch too cool mode instead. This should be more
effective at cooling and avoid drying out the air.