Completely refactor auto controller
Instead of mixing cooling and drying together in one hairball, have the
concept of entering a goal state and hitting the goal state. If we have
yet to hit a goal we just continue to try to reach it.

The temperature thresholds are given a higher priority than humidity. If
we exit the temperature range that will override dry mode.

This should:

1. Reduce code complexity

2. Avoid rapid state switching when near the cooling & humidity
   threshold

3. Allow drying the absence of a temperature threshold being hit