Fold dry mode in to fan mdoe
This treats dry mode as a special case of fan mode that allows for
greater outdoor humidity

Dry mode seems a lot more successful at changing the temperature than
the humidity. If we don't account for the temperature when activating
dry mode we run the risk of bouncing in to a temperature limit. This is
especially bad if we cool the room too quickly and need to take in a
bunch of wet outdoor air.

This also simplifies the code quite a bit and removes a bunch of
hardcoded humidity thresholds. This should make it easier to allow for
user-controlled humidity.