Add "yield AC" switch
If the AC is currently goal seeking it will be allowed to continue but
it will shut it off afterwards and not enter any new states.

This is intended to be hooked up to things like door sensors where we
don't want to aggressively switch AC states in response to normal
traffic.