Adjust AC target temperature while running
We can get stuck in a state where the AC is trying to hit a target
temperature on the wrong side of our target temperature. This
essentially wedges the whole system because:

1. We won't reach the user's desired temperature

2. The AC will stay on forever because it won't cross the temperature
   threshold

This adds a check that if our AC target temperature is on the wrong side
of our internal target then reset it to the other threshold. This should
catch the wedging states without constantly adjusting the target.